Integrate ticket check-in and marketplace features

+++ src/facets/FactoryFacet.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity 0.8.30;
 
-import {LibFactory} from "@host-it/libs/LibFactory.sol";
-import {TicketData, FullTicketData} from "@host-it-storage/FactoryStorage.sol";
-import {FeeType} from "@host-it-storage/MarketplaceStorage.sol";
+import {LibFactory} from "@ticket/libs/LibFactory.sol";
+import {TicketData, FullTicketData} from "@ticket-storage/FactoryStorage.sol";
+import {FeeType} from "@ticket-storage/MarketplaceStorage.sol";
 
 contract FactoryFacet {
     using LibFactory for *;
@@ -18,7 +18,7 @@ contract FactoryFacet {
         _ticketData._createTicket(_feeTypes, _fees);
     }
 
-    function updateTicket(TicketData calldata _ticketData, uint40 _ticketId) external {
+    function updateTicket(TicketData calldata _ticketData, uint56 _ticketId) external {
         _ticketData._updateTicket(_ticketId);
     }
 
@@ -26,15 +26,15 @@ contract FactoryFacet {
     //                               VIEW FUNCTIONS
     //////////////////////////////////////////////////////////////////////////*//
 
-    function ticketCount() public view returns (uint40) {
+    function ticketCount() public view returns (uint56) {
         return LibFactory._getTicketCount();
     }
 
-    function ticketExists(uint40 _ticketId) public view returns (bool) {
+    function ticketExists(uint56 _ticketId) public view returns (bool) {
         return _ticketId._ticketExists();
     }
 
-    function ticketData(uint40 _ticketId) public view returns (FullTicketData memory) {
+    function ticketData(uint56 _ticketId) public view returns (FullTicketData memory) {
         return _ticketId._getFullTicketData();
     }
 
@@ -42,7 +42,7 @@ contract FactoryFacet {
         return LibFactory._getAllFullTicketData();
     }
 
-    function adminTickets(address _ticketAdmin) public view returns (uint40[] memory) {
+    function adminTickets(address _ticketAdmin) public view returns (uint56[] memory) {
         return _ticketAdmin._getAdminTicketIds();
     }
 
@@ -58,15 +58,15 @@ contract FactoryFacet {
         return LibFactory._getHostItTicketHash();
     }
 
-    function ticketHash(uint40 _ticketId) public pure returns (bytes32) {
+    function ticketHash(uint56 _ticketId) public pure returns (bytes32) {
         return _ticketId._generateTicketHash();
     }
 
-    function mainAdminRole(uint40 _ticketId) public pure returns (uint256) {
+    function mainAdminRole(uint56 _ticketId) public pure returns (uint256) {
         return _ticketId._generateMainTicketAdminRole();
     }
 
-    function ticketAdminRole(uint40 _ticketId) public pure returns (uint256) {
+    function ticketAdminRole(uint56 _ticketId) public pure returns (uint256) {
         return _ticketId._generateTicketAdminRole();
     }
 }
